Rename test_vers to vers_utils
